+++ packages/cli/src/api/extractors/babel.ts
@@ -159,7 +159,10 @@ const extractor: ExtractorType = {
     const parserOptions = ctx.linguiConfig.extractorParserOptions
 
     // https://babeljs.io/docs/en/babel-parser#latest-ecmascript-features
-    const parserPlugins: ParserPlugin[] = []
+    const parserPlugins: ParserPlugin[] = [
+      "importAttributes", // stage3
+      "explicitResourceManagement", // stage3,
+    ]
 
     if (
       [/\.ts$/, /\.mts$/, /\.cts$/, /\.tsx$/].some((r) => filename.match(r))